refactor current subscription api call
