+++ src/widgets/profileSettings/components/accountManagement/currentSubscription/index.ts
@@ -1,1 +1,1 @@
-export * from './CurrentSubscription'
+export * from './ui/CurrentSubscription'
