+++ src/widgets/profileSettings/components/accountManagement/currentSubscription/model/subscriptionTypes.ts
@@ -1,0 +1,5 @@
+export type Subscription = {
+  createdAt?: string
+  expiresAt?: string
+  status?: string
+}
