+++ src/widgets/profileSettings/components/accountManagement/currentSubscription/ui/CurrentSubscription.module.scss
